# contracts/openapi/dashboard-v1.json
{
  "components": {
    "schemas": {
...
      "AuthenticationMode": {
        "enum": [
          "local",
          "oidc"
        ],
        "title": "AuthenticationMode",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequiredSession": {
        "properties": {
          "login_url": {
            "default": "/api/v1/auth/login",
            "title": "Login Url",
            "type": "string"
          },
          "provider_label": {
            "title": "Provider Label",
            "type": "string"
          },
          "status": {
            "const": "login_required",
            "default": "login_required",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "provider_label"
        ],
        "title": "LoginRequiredSession",
        "type": "object"
      },
      "PairingRequest": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "PairingRequest",
        "type": "object"
      },
...
      "PublicSessionStatus": {
        "properties": {
          "status": {
            "const": "pairing_required",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "PublicSessionStatus",
        "type": "object"
      },
...
      "SessionResponse": {
        "properties": {
          "authentication_mode": {
            "$ref": "#/components/schemas/AuthenticationMode"
          },
          "csrf_token": {
            "title": "Csrf Token",
            "type": "string"
          },
          "idle_expires_at": {
            "format": "date-time",
            "title": "Idle Expires At",
            "type": "string"
          },
          "status": {
            "const": "authenticated",
            "default": "authenticated",
            "title": "Status",
            "type": "string"
          },
          "user": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "workspace": {
            "$ref": "#/components/schemas/WorkspaceSummary"
          },
          "workspaces": {
            "items": {
              "$ref": "#/components/schemas/WorkspaceSummary"
            },
            "title": "Workspaces",
            "type": "array"
          }
        },
        "required": [
          "csrf_token",
          "idle_expires_at",
          "authentication_mode",
          "user",
          "workspace",
          "workspaces"
        ],
        "title": "SessionResponse",
        "type": "object"
      },
...
      "UserSummary": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "display_name"
        ],
        "title": "UserSummary",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkspaceRole": {
        "enum": [
          "owner",
          "member"
        ],
        "title": "WorkspaceRole",
        "type": "string"
      },
      "WorkspaceSummary": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/WorkspaceRole"
          }
        },
        "required": [
          "id",
          "name",
          "role"
        ],
        "title": "WorkspaceSummary",
        "type": "object"
      },
      "WorkspaceSwitchRequest": {
        "additionalProperties": false,
        "properties": {
          "workspace_id": {
            "format": "uuid",
            "title": "Workspace Id",
            "type": "string"
          }
        },
        "required": [
          "workspace_id"
        ],
        "title": "WorkspaceSwitchRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Parserium Collector",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/session": {
      "get": {
        "operationId": "get_session_api_v1_session_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "parserium_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parserium Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session",
        "tags": [
          "session"
        ]
      }
    },
    "/api/v1/session/logout": {
      "post": {
        "operationId": "logout_session_api_v1_session_logout_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Parserium-CSRF",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Parserium-Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "parserium_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parserium Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Logout Session",
        "tags": [
          "session"
        ]
      }
    },
    "/api/v1/session/pair": {
      "post": {
        "operationId": "pair_session_api_v1_session_pair_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pair Session",
        "tags": [
          "session"
        ]
      }
    },
    "/api/v1/session/status": {
      "get": {
        "operationId": "get_session_status_api_v1_session_status_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "parserium_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parserium Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/PublicSessionStatus"
                    },
                    {
                      "$ref": "#/components/schemas/LoginRequiredSession"
                    },
                    {
                      "$ref": "#/components/schemas/SessionResponse"
                    }
                  ],
                  "title": "Response Get Session Status Api V1 Session Status Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session Status",
        "tags": [
          "session"
        ]
      }
    },
    "/api/v1/session/workspace": {
      "post": {
        "operationId": "switch_workspace_api_v1_session_workspace_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Parserium-CSRF",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Parserium-Csrf"
            }
          },
          {
            "in": "cookie",
            "name": "parserium_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parserium Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkspaceSwitchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Switch Workspace",
        "tags": [
          "session"
        ]
      }
    }
  }
}